// src/app/services/models/student-dto-to-return.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface StudentDtoToReturn {
  id:number,
  email: string;
  firstName?: string;
  lastName?: string;
  phone?: string;
  schoolYear?: string;
  sexe?: 'Men' | 'Women';
  imageName?:string;
  createdDate?:string;
  birthday?:string;

}