// src/app/services/models/professor-to-return.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProfessorToReturn {
  id?:number,
  departement?: 'BackendDevelopemet' | 'FullStackDevelopement' | 'FrontDevelopement' | 'AI_MachineLearning';
  email?: string;
  firstName?: string;
  lastName?: string;
  phone?: string;
  sexe?: 'Men' | 'Women';
  imageName?:string;
  createdDate?:string;
  birthday?:string;


}